16/08/22
On this day, added two new html files, one is for the delivery mode. In this HTML file it includes a form where the user can enter their name, phone number and address.

While the second HTML file is the pickup mode, where in this file it only includes a form where the user can enter only their name