17/07/22
Added a new variable called pickup where if the button found in the delivery form is clicked it will assign the  variable to false. However if the button found within the pickup form is clicked it will assign the variable named "pickup" to true.

This is done so that I can use this variable in the future especially in the menu function when displaying the user's information and the total.