19/09/22
On this day I have added a two functions. The first function is to change the total whenever the user adds or deletes an item from the shopping cart.

And the second function, is that when the user changes the quantities of coffee they want to buy, the program needs to get the input that they have entered and show that in the total. I also have added some error trapping to this function, for example if the input is less than 1 or less than 10, it will just enter the value that they want as one. Since it is already part of the shopping cart