20/09/22 - 22/09/22
On these days, I have created the final functions and final sections for my program.

On these days, I have created a function to show the final chosen products in the order information section. I have named this function as "orderInformation()", where in this function, it will just retrieve the items or the products that is found inside the shopping cart, and use that items and show them to the order informations section.  Please see log file for more details.

The next function I have coded on these days, is then going to be the "orderInfoFInal()" function, in this function it is coding for how the section of the final order information is going to look like. For example, it will determine if the user has chosen delivery or not, and if the user has chosen delivery, a $5 delivery charge will be added to the total, and if the user did choose pickup it will just say "no delivery fee". And other more aspects that is specified within the assessment requirements.